utils/sdxl_accuracy/requirements.txt
requests
Pillow
torch
open-clip-torch
scipy
loguru